+++ lib/ex_unit_json/config.ex
@@ -33,7 +33,7 @@ defmodule ExUnitJSON.Config do
           group_by_error: boolean()
         ]
 
-  @valid_options [:summary_only, :failures_only, :first_failure, :filter_out, :output, :compact, :group_by_error]
+  @valid_options [:summary_only, :failures_only, :first_failure, :filter_out, :output, :compact, :group_by_error, :quiet]
 
   @doc """
   Gets options from Application environment.